* cleaned up code for Mac OS from iOS specific code * cleaned clang warnings

+++ iOS/3.4/pubnub/libs/PubNub/Data/Channels/Presence/PNPresenceEvent.m
@@ -122,8 +122,9 @@ struct PNPresenceEventDataKeysStruct PNPresenceEventDataKeys = {
     }
 
 
-    return [NSString stringWithFormat:@"%@ \nEVENT: %@\nUSER IDENTIFIER: %@\nDATE: %@\nOCCUPANCY: %d\nCHANNEL: %@",
-                    NSStringFromClass([self class]), action, self.uuid, self.date, self.occupancy, self.channel];
+    return [NSString stringWithFormat:@"%@ \nEVENT: %@\nUSER IDENTIFIER: %@\nDATE: %@\nOCCUPANCY: %ld\nCHANNEL: %@",
+                    NSStringFromClass([self class]), action, self.uuid, self.date, (unsigned long)self.occupancy,
+                    self.channel];
 }
 
 #pragma mark -
